pyproject.toml
[project]
name = "mcp-data-check"
version = "0.1.0"
description = "Evaluate MCP server accuracy against known questions and answers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_data_check"]

[project.scripts]
mcp-data-check = "mcp_data_check.run_eval:main"
